+++ core/module/page/view/edit/edit.js.php
@@ -15,7 +15,7 @@
 /**
  * Confirmation de suppression
  */
-$("#pageEditDelete").on("click", function() {
+ $("#pageEditDelete").on("click", function() {
 	var _this = $(this);
 	return core.confirm("Confirmez-vous la suppression de cette page ?", function() {
 		$(location).attr("href", _this.attr("href"));
@@ -239,10 +239,9 @@ $( document ).ready(function() {
 	 */
 		 if ($("#pageEditExtraPosition").val() == 1 ) {
 			var positionDOM = $("#pageEditPosition");
-			positionDOM.empty().append(
-				$("<option>").val(0).text("Ne pas afficher"),
-				$("<option>").val(1).text("Au début")
-			);
+			var positionInitial = <?php echo $this->getData(['page',$this->getUrl(2),"position"]); ?>;
+			buildPagesList(true);
+			$("#pageEditPosition").val(positionInitial);
 		}
 
 });
@@ -523,14 +522,9 @@ pageTypeMenuDOM.on("change", function() {
 
 $("#pageEditExtraPosition").on("change", function() {
 	if ($("#pageEditExtraPosition").val() == 1 ) {
-
-		var positionDOM = $("#pageEditPosition");
-		positionDOM.empty().append(
-			$("<option>").val(0).text("Ne pas afficher"),
-			$("<option>").val(1).text("Au début")
-		);
+		buildPagesList(true);
 	} else {
-		getPages();
+		buildPagesList(false);
 		//$("#pageEditParentPageId").trigger("change");
 	}
 });
@@ -546,14 +540,14 @@ $("#pageEditModuleConfig").on("click", function() {
  * Affiche les pages en fonction de la page parent dans le choix de la position
  */
 $("#pageEditParentPageId").on("change", function() {
-	getPages();
+	buildPagesList(false);
 }).trigger("change");
 
 /**
  * Construit un select contenant la liste des pages du site.
  */
 
-function getPages() {
+function buildPagesList(extraPosition) {
 	var hierarchy = <?php echo json_encode($this->getHierarchy()); ?>;
 	var pages = <?php echo json_encode($this->getData(['page'])); ?>;
 	var positionInitial = <?php echo $this->getData(['page',$this->getUrl(2),"position"]); ?>;
@@ -566,11 +560,12 @@ function getPages() {
 	var positionSelected = 0;
 	var positionPrevious = 1;
 
-	// Aucune page parent selectionnée
+	// Aucune page parent sélectionnée
 	if(parentSelected === "") {
 		// Liste des pages sans parents
 		for(var key in hierarchy) {
-			if(hierarchy.hasOwnProperty(key)) {
+			if(hierarchy.hasOwnProperty(key)
+			  && extraPosition === pages[key].extraPosition) {
 				// Sélectionne la page avant s'il s'agit de la page courante
 				if(key === "<?php echo $this->getUrl(2); ?>") {
 					positionSelected = positionPrevious;
@@ -590,7 +585,7 @@ function getPages() {
 			positionSelected = 0;
 		}
 	}
-	// Un page parent est selectionnée
+	// Une page parent est sélectionnée
 	else {
 		// Liste des pages enfants de la page parent
 		for(var i = 0; i < hierarchy[parentSelected].length; i++) {